chore(i18n): move localization files from public to src folder
