+++ src/i18n.ts
@@ -2,8 +2,8 @@ import { initReactI18next } from 'react-i18next'
 
 import i18n from 'i18next'
 
-import { resources } from '../public/locales'
 import { I18N_DEFAULT_LANGUAGE, I18N_DEFAULT_NS } from './constants'
+import { resources } from './locales'
 
 i18n.use(initReactI18next).init({
 	resources: resources,
